Move assignable func impl into thema package
